Create README.md
Each group will evaluate one assigned category in Dashboard Admin area in WordPress. Each group will respond the the following:


What is the category for?
How should it be used?
How will you use it?
What are the recommended settings?
What are the sub-categories?
How are the sub-categories used?
Show others how to use the category and sub-category if needed?
Collaborate with each other to create a short -- 4-5 slide presentation that includes a title slide with your group name and member names, your research and response to the above questions, and a tutorial/recommendation for users.